% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/CP_model_inv.R
\name{CP_model_inv}
\alias{CP_model_inv}
\title{Inverse critical power modelling.}
\usage{
CP_model_inv(P, tsec)
}
\arguments{
\item{P}{numeric; maximal mean power values for time periods given in the
tsec argument.}

\item{tsec}{numeric; time values (seconds) that correspond to elements in P.}
}
\value{
a named vector of parameters (CP and W').
}
\description{
Generate critical power parameters via a linearised inverse time model.
Mainly useful when the length of inputs is <= 2, making
\code{\link[stats]{nls}} inappropriate.
}

