% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{cycling_data}
\alias{cycling_data}
\title{Example cycling data.}
\format{An object of class c("cycleRdata", "data.frame").
\describe{
\item{timer.s}{an ongoing timer (seconds). Stoppages are not recorded per se,
but rather represented as breaks in the continuity of the timer.}
\item{timer.min}{as above, but in units of minutes.}
\item{timestamp}{"POSIXct" values, describing the actual time of day.}
\item{delta.t}{delta time values (seconds). Obtained via base::diff(timer.s).}
\item{lat}{latitude values (units: degrees).}
\item{lng}{longitude values (units: degrees).}
\item{distance.km}{cumulative distance (kilometres).}
\item{speed.kmh}{speed in kilometres per hour.}
\item{elevation.m}{altitude in metres.}
\item{delta.elev}{delta elevation (metres). Obtained via base::diff(elevation.m).}
\item{VAM}{"vertical ascent metres per second".}
\item{power.W}{power readings (watts).}
\item{power.smooth.W}{an exponentially-weighted 25-second moving average of power.W values.}
\item{work.J}{cumulative work (joules).}
\item{Wexp.kJ}{W' expended in units of kilojoules. See ?Wbal and references therein.}
\item{lap}{a numeric vector of lap "levels". Will only have values > 1 if lap data is available.}
}}
\usage{
cycling_data
}
\description{
Formatted cycling data from a Garmin head-unit. imported via
cycleRtools::read_fit("file_path", format = TRUE).
}
\keyword{datasets}

