% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dist_plot.R
\name{dist_plot}
\alias{dist_plot}
\title{Time distribution plot.}
\usage{
dist_plot(data, column = "power.W", binwidth = 10, zbounds = NULL,
  xlab = NULL, xlim = NULL, ylim = NULL)
}
\arguments{
\item{data}{a \strong{formatted} dataset produced by \code{read*()}.}

\item{column}{column in \code{data} giving the values of interest. Needn't be
quoted.}

\item{binwidth}{how should values in \code{column} be binned? E.g.
\code{bindiwdth = 10} will create 10 watt bins if \code{column} is power
data.}

\item{zbounds}{optional; a numeric vector of zone boundaries.}

\item{xlab}{x axis label.}

\item{xlim}{numeric vector of length 2; x axis limits.}

\item{ylim}{numeric vector of length 2; y axis limits.}
}
\value{
a ggplot object.
}
\description{
Display the time distribution of values within a dataset. The distribution
can also be partitioned into zones if the \code{zbounds} argument is not
\code{NULL}.
}
\examples{
data(cycling_data)
dist_plot(
 data = cycling_data,
 column = power.W,
 binwidth = 10,
 zbounds = c(100, 200, 300),
 xlim = c(110, 500)
)
}

