% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/interval_detect.R
\name{interval_detect}
\alias{interval_detect}
\title{Detect Intervals in a Ride.}
\usage{
interval_detect(data, sections, plot = FALSE, ...)
}
\arguments{
\item{data}{a \strong{formatted} dataset produced by \code{read*()}.}

\item{sections}{how many sections should be identified? see
\strong{Description}.}

\item{plot}{logical; if TRUE, graphically displays the resultant sections.}

\item{...}{graphical parameters to be passed to \code{par()}. Not used if
  \code{plot = FALSE}.}
}
\value{
if \code{plot = TRUE} nothing is returned. If \code{plot = FALSE}
  (default) a vector of section "levels" is returned.
}
\description{
Often a ride will contain intervals/efforts that are not in any way marked in
the device data (e.g. as "laps"). Using changepoint analysis, it is possible
to retrospectively identify these efforts. This is contingent on supplying
the number of changepoints to the underlying algorithm, simplified here as a
\code{sections} argument. For example, if there are two efforts amidst a
ride, this means we are looking to identify 5 \emph{sections} (i.e.
neutral-effort-neutral-effort-neutral). See \strong{Examples}. Requires
package \code{"changepoint"}.
}
\examples{
# "interval_data" is a ride data set, which involved two 10 minute efforts
# during an otherwise steady ride.
data(interval_data)
interval_data$interval <- interval_detect(interval_data, 5, FALSE)
# Two efforts = 5 sections.
# Were the two 10 minute efforts properly identified?
tapply(interval_data$delta.t, interval_data$interval, sum) / 60
# Plot
interval_detect(interval_data, 5, TRUE)
}

