% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mmv.R
\name{mmv}
\alias{mmv}
\title{Maximal mean values.}
\usage{
mmv(data, column, windows, delta = NULL, verbose = TRUE, .uniform = FALSE,
  .string = FALSE)
}
\arguments{
\item{data}{a \strong{formatted} dataset produced by \code{read*()}.}

\item{column}{column in \code{data} giving the values of interest. Needn't be
quoted.}

\item{windows}{window size(s) for which to generate best averages, given in
seconds.}

\item{delta}{the incremement with which to initially make the data uniform.
If NULL (default), the most appropriate value is estimated.}

\item{verbose}{should messages be printed to the console?}

\item{.uniform}{are the rows of the data already uniform?}

\item{.string}{are column name arguments given as character strings? A
backdoor around non-standard evaluation. Mainly for internal use.}
}
\value{
a matrix object with two rows: 1) best mean value(s) and 2) the time
  at which that value was recorded
}
\description{
Calculate maximal mean values for specified time periods.
}
\examples{
data(cycling_data)
# Generate best powers for 5, 10 and 20 minutes.
t_sec <- c(5, 10, 20) * 60
x <- mmv(cycling_data, power.W, t_sec)
# Show when those values were recorded in minutes
x[2, ] / 60

}
\seealso{
For a more generic and efficient version of this function, see
  \code{\link{mmv2}}
}

