% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/NP.R
\name{NP}
\alias{NP}
\title{Normalised power.}
\usage{
NP(data, delta = NULL, verbose = TRUE)
}
\arguments{
\item{data}{a \strong{formatted} dataset produced by \code{read*()}.}

\item{delta}{optional; the sampling frequency of data (in seconds per
sample).}

\item{verbose}{should messages be printed to the console?}
}
\value{
a Normalised Power value.
}
\description{
Calculate a Normalised Power value. Normalised Power is a registered
trademark of Peaksware Inc.
}
\examples{
data(cycling_data)
NP(cycling_data)
}

