% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Pt_prof}
\alias{Pt_prof}
\title{Power-time profile.}
\format{a \code{data.frame} with two columns: time (seconds) and power
  (Watts), respectively.}
\usage{
Pt_prof
}
\description{
An example power profile; i.e. best mean powers for periods of 30 seconds
through to 1 hour, in increments of 10 seconds.
}
\keyword{datasets}

