% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/pwr_TRIMP.R
\name{pwr_TRIMP}
\alias{pwr_TRIMP}
\title{Power-Based TRaining IMPulse.}
\usage{
pwr_TRIMP(data, CP)
}
\arguments{
\item{data}{a \strong{formatted} dataset produced by \code{read*()}.}

\item{CP}{a critical power value (watts).}
}
\value{
a normalised TRIMP score.
}
\description{
Calculate a \emph{normalised} TRIMP value using power data. This is a
power-based adaptation of Bannister's TRIMP, whereby critical power (CP) is
assumed to represent 90% of heart rate ratio (HRR). The final TRIMP score is
normalised to the score associated with one-hour's riding at CP, to aid
interpretation.
}
\references{
Morton, R.H., Fitz-Clarke, J.R., Banister, E.W., 1990. Modeling
  human performance in running. Journal of Applied Physiology 69, 1171-1177.
}

