% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reset.R
\name{reset}
\alias{reset}
\title{Reset a dataset or vector.}
\usage{
reset(x)
}
\arguments{
\item{x}{a numeric vector or formatted cycling dataset (i.e. class \code{"cycleRdata"}).}
}
\value{
either a data frame or vector, depending on the class of \code{x}.
}
\description{
Subtracts the first element from all other elements.
}
\details{
if x is a formatted dataset from a \code{read*} function, all the
  columns are reset as appropriate. This can be useful after subsetting a
  ride dataset, for example. Otherwise, this is a simple shortcut for \code{x
  - x[[1]]}.
}
\examples{
data(cycling_data)
# Remove first minute of data and reset.
data_raw <- cycling_data[cycling_data$timer.s > 60, ]
data_reset <- reset(data_raw)
# Compare...
data_raw$distance.km[[1]]
data_reset$distance.km[[1]]

}

