% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{ride_examples}
\alias{intervaldata}
\alias{ride_examples}
\alias{ridedata}
\title{Example cycling data.}
\format{An object of class \code{c("cycleRdata", "data.frame")}, and
  additional attributes of \code{CP = 300} & \code{sRPE = 7}. The latter are
  used by several methods in this package. See \code{\link{cycleRdata}} for a
  description of columns.}
\usage{
ridedata

intervaldata
}
\description{
Formatted cycling data from a Garmin head-unit. Imported via
\code{read_fit("file_path", format = TRUE, CP = 310, sRPE = 7)}.

\code{"ridedata"} is a typical group ride. \code{"intervaldata"} is a session
(of sorts) that included two efforts and a cafe stop. The latter is included
to demonstrate the use of \code{\link{interval_detect}}.
}
\seealso{
\code{\link{cycleRdata}}.
}
\keyword{datasets}

