% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ride_time.R
\name{ride_time}
\alias{ride_time}
\title{Calculate ride time.}
\usage{
ride_time(x, delta = NULL)
}
\arguments{
\item{x}{a vector of time values.}

\item{delta}{numeric; the typical interval between time values, if \code{NULL} a
best estimate is used.}
}
\value{
ride time in the same units as x.
}
\description{
A simple function for calculating ride time as opposed to elapsed time.
}
\examples{
t_sec <- c(1:20, 50:70)
# elapsed time
max(t_sec)
# ride time.
ride_time(t_sec)

}

