% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{rollmean_nunif}
\alias{rollmean_nunif}
\title{Rolling mean for nonuniform data.}
\usage{
rollmean_nunif(x, t, window)
}
\arguments{
\item{x}{numeric vector of values to be rolled.}

\item{t}{numeric vector of time values corresponding to elements in \code{x}.}

\item{window}{size of the window in terms of \code{t}. E.g. \code{30} (seconds).}
}
\description{
Produce a rolling average for data sampled at non-uniform time intervals.
}

