% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smth_plot.R
\name{smth_plot}
\alias{smth_plot}
\title{Smoothed data plot.}
\usage{
smth_plot(data, x = "timer.s", yraw = "power.W", ysmth = "power.smooth.W",
  colour = "lap", ..., .string = FALSE)
}
\arguments{
\item{data}{the dataset to be used.}

\item{x}{column identifier for the x axis data.}

\item{yraw}{column identifier for the (underlying) raw data.}

\item{ysmth}{column identifier for the smoothed data.}

\item{colour}{level identifier in \code{data} by which to colour lines. Or a
colour name. Or simply a nonexistent column name (equivalent to "black").}

\item{...}{further arguments to be passed to \code{plot()}.}

\item{.string}{are column name arguments given as character strings? A
  backdoor around non-standard evaluation. Mainly for internal use.}
}
\description{
Create a plot with both raw and smoothed data lines.
}
\examples{
data(cycling_data)
# Create some artificial laps.
cycling_data$lap <- floor(seq(from = 1, to = 5, along.with = cycling_data[, 1]))
# Plot:
smth_plot(cycling_data, timer.s, power.W, power.smooth.W, colour = "lap")
# Plot with a single blue line:
smth_plot(cycling_data, colour = "blue")  # Default arguments.
}

