% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/TSS.R
\name{TSS}
\alias{TSS}
\title{Training stress score.}
\usage{
TSS(data, threshold, verbose = TRUE)
}
\arguments{
\item{data}{a \strong{formatted} dataset produced by \code{read*()}.}

\item{threshold}{a threshold value - e.g. FTP.}

\item{verbose}{should messages be printed to the console?}
}
\value{
a TSS score.
}
\description{
Calculate a Training Stress Score (TSS). TSS is a registered trademark of
Peaksware Inc.
}

