% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uniform.R
\name{uniform}
\alias{uniform}
\title{Create a uniform data.frame.}
\usage{
uniform(data, delta = NULL, verbose = TRUE, .return_delta = FALSE)
}
\arguments{
\item{data}{a data.frame or matrix object to make uniform. This is done on
the basis of the first column.}

\item{delta}{the incremement by which to make the data uniform. If
\code{NULL} (default) the most appropriate value is estimated.}

\item{verbose}{should messages be printed to the console?}

\item{.return_delta}{if TRUE the delta value is assigned to the calling
environment. Intended for internal use.}
}
\value{
A uniform data frame. An index column is appended so that original
  data values can be extracted - e.g. after a rolling operation.
}
\description{
Returns a data frame with rows that are uniformly incremented along the first
column. Useful for rolling operations, whereby a rolling time window is of
interest.
}
\examples{
t_sec <- c(1:10, 20:40, 50)            # Discontinuous timer values.
pwr   <- runif(length(t_sec), 0, 400)  # Some power values.
x     <- data.frame(t_sec, pwr)
uniform(x)

}
\seealso{
source code for \code{\link{rollmean_smth}}
}

