% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Wbal_plots.R
\name{Wbal_plots}
\alias{Wbal_plots}
\title{W' balance plots.}
\usage{
Wbal_plots(data, x = 1, n = c(1, 2, 3), xlab = NULL, xlim = NULL,
  CP = NULL, laps = FALSE, ...)
}
\arguments{
\item{data}{a \strong{formatted} dataset produced by \code{read*()}.}

\item{x}{numeric; \code{1} will plot against time (sec); \code{2} will plot
against time (minutes); and \code{3} will plot against distance.}

\item{n}{plots to be created (see details).}

\item{xlab}{character; x axis label for bottom plot.}

\item{xlim}{given in terms of \code{x}.}

\item{CP}{a value for critical power annotation.}

\item{laps}{logical; should laps be seperately coloured?}

\item{...}{graphical parameters to be passed to \code{par()}.}
}
\value{
a variable number of plots.
}
\description{
Generate three plots that effectively summarise a whole cycling dataset.
}
\details{
The \code{n} argument describes plot options such that: \enumerate{\item
plots W' balance (kJ). \item plots power data (W). \item plots an elevation
profile (m).} These options can be concatenated to produce a stack of plots
as desired.
}
\examples{
data(cycling_data)
Wbal_plots(cycling_data, x = 2, n = c(1, 2, 3), CP = 300)
# Show just W' balance.
Wbal_plots(cycling_data, x = 2, n = 1)
# Elevation profile on top
Wbal_plots(cycling_data, x = 2, n = c(3, 1))
# Zoom all plots to 20-40 km.
Wbal_plots(cycling_data, x = 3, CP = 300, xlim = c(20, 40), xaxs = "i")

}
\seealso{
\code{\link{Wbal}}.
}

