% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zdist_plot.R
\name{zdist_plot}
\alias{zdist_plot}
\title{Zone-time distribution plot.}
\usage{
zdist_plot(data, column = "power.W", binwidth = 10, zbounds = NULL,
  .string = FALSE, ...)
}
\arguments{
\item{data}{a \strong{formatted} dataset produced by \code{read*()}.}

\item{column}{column in \code{data} giving the values of interest. Needn't be
quoted.}

\item{binwidth}{how should values in \code{column} be binned? E.g.
\code{bindiwdth = 10} will create 10 watt bins if \code{column} is power
data.}

\item{zbounds}{optional; a numeric vector of zone boundaries.}

\item{.string}{are column name arguments given as character strings? A
backdoor around non-standard evaluation. Mainly for internal use.}

\item{...}{arguments to be passed to \code{barplot()} and/or graphical
  parameters.}
}
\value{
nothing; a plot is sent to the current graphics device.
}
\description{
Display the time distribution of values within a dataset. The distribution
can also be partitioned into zones if the \code{zbounds} argument is not
\code{NULL}.
}
\examples{
data(cycling_data)
zdist_plot(
 data = cycling_data,
 column = power.W,
 binwidth = 10,
 zbounds = c(100, 200, 300),
 xlim = c(110, 500)
)
}

