% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convert_time.R
\name{convert_time}
\alias{convert_from_time}
\alias{convert_time}
\alias{convert_to_time}
\title{Reformat time.}
\usage{
convert_from_time(x)

convert_to_time(x)
}
\arguments{
\item{x}{either a character string of the form "HH:MM:SS" ("HH" is optional)
or a seconds value; can accept vectors.}
}
\value{
seconds value(s) for \emph{from}, and "HH:MM:SS" character string(s)
  for \emph{to}.
}
\description{
Functions perform interconversion between "HH:MM:SS" format and seconds.
}

