% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/convert_time.R
\name{convert_time}
\alias{convert_time}
\title{Reformat time.}
\usage{
convert_time(x)
}
\arguments{
\item{x}{character string; a time in (HH:)MM:SS format. HH is optional.}
}
\value{
a numeric time value in seconds.
}
\description{
Convert time from "HH:MM:SS" format to seconds.
}
\examples{
convert_time("3:21:05")       # Simple example.
convert_time("03:30:10.5")    # With decimal seconds.
convert_time("19:30.1")       # Without hours.
convert_time("19:30.1") / 60  # In minutes.
}

