% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{diff_section}
\alias{diff_section}
\title{Section data according to breaks.}
\usage{
diff_section(x)
}
\arguments{
\item{x}{a numeric vector (e.g. a timer column) that increments uniformly.
When there is a \strong{significant} break in this uniformity, a new
section is created, and so forth.}
}
\value{
a vector of the same shape as x.
}
\description{
Generates a vector of "section" values/levels according to differences in
the supplied vector. The function simply rolls over \code{x}, incrementing
the return vector every time there is a significant break in the pattern of
differences between adjacent elements of \code{x}. In practical terms, if
\code{x} is a series of timestamp values (see example), every time there is
a significant break in the timer (> 10 sec currently), the return vector is
incremented by 1.
}
\examples{
t_sec <- c(1:10, 40:60, 100:150)       # Discontinuous timer values.
pwr   <- runif(length(t_sec), 0, 400)  # Some power values.
x     <- data.frame(t_sec, pwr)
# Generate section levels.
x$section <- diff_section(x$t_sec)
print(x)

}

