% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elevation_correct.R
\name{download_elev_data}
\alias{download_elev_data}
\title{Download geographical elevation data.}
\usage{
download_elev_data(country = "all")
}
\arguments{
\item{country}{character string; the ISO3 country code (see
\code{raster::getData("ISO3")}) for which to download the data. If "all",
then all available data is downloaded - this may take some time.}
}
\value{
nothing, files are downloaded to the working directory.
}
\description{
Downloads elevation data files to the working directory for use with
\code{\link{elevation_correct}}. Requires package \code{raster} to be
installed.
}
\seealso{
\code{\link{elevation_correct}}.
}

