% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{mmv2}
\alias{mmv2}
\title{Efficient maximal mean values.}
\usage{
mmv2(x, pds)
}
\arguments{
\item{x}{a numeric vector of values.}

\item{pds}{window size(s) (in element units) for which to
  generate maximum mean values.}
}
\value{
a vector of \code{length(pds)}.
}
\description{
A more efficient implementation of \code{\link{mmv}}. Simply takes
a vector (\code{x}) of values and rolls over them element wise by windows
defined in \code{pds}. Returns a vector of maximum mean values for each
window size.
}

