% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Ptmodels_S3.R
\name{predict.Ptmodels}
\alias{predict.Ptmodels}
\title{Predict Power or Time}
\usage{
\method{predict}{Ptmodels}(object, x, y = "P", ...)
}
\arguments{
\item{object}{an object of class "Ptmodels".}

\item{x}{the value for which to make a prediction, see below.}

\item{y}{the type of variable to predict: "P" will produce a power
prediction, and hence assume \code{x} is a time value; and "tsec" vice
versa.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a named numeric vector of predicted values. Names correspond to their
  respective models.
}
\description{
Given a Ptmodels \code{object}, the predict.Ptmodels will produce a named
numeric vector of either time (seconds) or power (watts) values according to
the \code{x} and \code{y} arguments
}

