% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Pt_prof}
\alias{Pt_prof}
\title{An example Power-time profile}
\format{\preformatted{ num [1:358(1d)] 993 896 870 849 811 ...
 - attr(*, "dimnames")=List of 1
  ..$ : chr [1:358] "30" "40" "50" "60" ...
}}
\usage{
Pt_prof
}
\description{
A named numeric vector of best power values, where names are the
corresponding time periods (seconds).
}
\keyword{datasets}

