% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Pt_prof}
\alias{Pt_prof}
\title{An example Power-time profile}
\format{An object of class \code{array} of length 358.}
\usage{
Pt_prof
}
\description{
A named numeric vector of best power values, where names are the
corresponding time periods (seconds).
}
\keyword{datasets}

