% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/read.R, R/read_fit.R, R/read_pwx.R, R/read_pwx2.R, R/read_srm.R, R/read_tcx2.R
\name{read}
\alias{read}
\alias{read_fit}
\alias{read_pwx}
\alias{read_pwx2}
\alias{read_srm}
\alias{read_tcx2}
\title{Read cycling device data.}
\usage{
read(file, format = TRUE, .list = FALSE)

read_fit(file, format = TRUE)

read_pwx(file, format = TRUE, .list = FALSE)

read_pwx2(file, format = TRUE, .list = FALSE)

read_srm(file, format = TRUE)

read_tcx2(file, format = TRUE, .list = FALSE)
}
\arguments{
\item{file}{character; path to the file.}

\item{format}{logical; should data be formatted?}

\item{.list}{should data be returned as a list? Useful when data columns are
  of unequal length, as can sometimes be the case with \code{read_pwx} and
  \code{read_tcx}.}
}
\value{
A data frame or list, according to the \code{.list} argument.
}
\description{
Read data from a cycling head unit into the R environment; optionally
formatting it for use with other functions in this package.
}
\details{
Returns a data frame with all data parsed from the file; with an additional
lap (factor) column appended in the case of \code{read_fit()}.

Note that most functions within this package depend on imported data being
formatted; i.e. \code{read*("file_path", format = TRUE)}. Hence, unless the
raw data is of particular interest and/or the user wants to process it
manually, the format argument should be TRUE (default). When working with a
formatted dataset, do not change existing column names. The formatted data
structure is described in detail in \link{cycling_data}.

Garmin .fit file data is parsed with the java command line tool provided in
the \href{http://www.thisisant.com/resources/fit}{FIT SDK}. The latest source
code and licensing information can be found at the previous link.

SRM device files (.srm) are also parsed at the command line, provided
\href{http://www.zuto.de/project/srmio/}{Rainer Clasen's srmio library} is
installed and available. The associated GitHub repo' can be found
\href{https://github.com/rclasen/srmio}{here}.
}
\section{Functions}{
\itemize{
\item \code{read}: A wrapper for read_* functions that chooses the appropriate
  function based on file extension.

\item \code{read_fit}: Read a Garmin (Ltd) device .fit file. This invokes
  \code{\link[base]{system2}} to execute the FitCSVTool.jar command line tool
  (see \href{http://www.thisisant.com/resources/fit}{FIT SDK}). Hence, this
  function requires that Java (JRE/JDK) binaries be on the system path.

\item \code{read_pwx}: Read a Training Peaks .pwx file. Requires the "XML"
package to be installed. \code{\link{read_pwx2}} is a faster alternative
that uses system commands.

\item \code{read_pwx2}: Read a Training Peaks .pwx file. Requires the command
line tool "xml2" to be installed.

\item \code{read_srm}: Read an SRM (.srm) file. This requires
\href{http://www.zuto.de/project/srmio/}{Rainer Clasen's srmio library} to
be installed and on the system path.

\item \code{read_tcx2}: Read a Garmin .tcx file. Fast and dirty command line
parsing. Requires the command line tool "xml2" to be installed.
}}

