% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/rollmean_smth.R
\name{rollmean_smth}
\alias{rollmean_smth}
\title{Smooth a column of a dataset.}
\usage{
rollmean_smth(data, column, smth.pd, ema = FALSE, delta = NULL,
  verbose = TRUE, .uniform = FALSE, .string = FALSE)
}
\arguments{
\item{data}{a data.frame or matrix object to use for smoothing. The data is
made \code{\link{uniform}} on the basis of the first column.}

\item{column}{the column name of the data to be smoothed, needn't be quoted.}

\item{smth.pd}{numeric; the time period over which to smooth (seconds).}

\item{ema}{should the moving average be exponentially weighted?}

\item{delta}{the incremement with which to initially make the data uniform.
If NULL (default) a best estimate is used.}

\item{verbose}{should messages be printed to the console?}

\item{.uniform}{are the rows of the data already uniform?}

\item{.string}{are column name arguments given as character strings? A
  backdoor around non-standard evaluation. Mainly for internal use.}
}
\value{
a vector of the same length as the \code{column}.
}
\description{
Smooth data with a right-aligned (zero-padded) rolling average. This is
intended for time-based smoothing - e.g. a 30-second moving average. Hence,
as rolling operations are performed row-wise in in the interest of
efficiency, the data must first be made time-uniform by row. Hence, this is
done via \code{\link{uniform}} by the first column of the data; thus allowing
the \code{column} to be smoothed.
}
\examples{
\dontrun{
data(cycling_data)
# Smooth power data with a 30 second moving average.
rollmean_smth(cycling_data, power.W, 30)
# Or alternatively, use an exponentially weighted moving average.
rollmean_smth(cycling_data, power.W, 30, ema = TRUE)
}
}

