% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/smth_plot.R
\name{smth_plot}
\alias{smth_plot}
\title{Smoothed data plot.}
\usage{
smth_plot(data, x = "timer.s", yraw = "power.W", ysmth = "power.smooth.W",
  xlim = NULL, ...)
}
\arguments{
\item{data}{the dataset to be used.}

\item{x}{column identifier for the x axis data.}

\item{yraw}{column identifier for the (underlying) raw data.}

\item{ysmth}{column identifier for the smoothed data.}

\item{xlim}{optional xlim values to be passed to \code{plot()}.}

\item{...}{further arguments to be passed to \code{plot()}.}
}
\description{
Create a plot with both raw and smoothed data lines. Convenience wrapper for
\link[graphics]{plot} whereby two plots are overaid with \code{par(new =
TRUE)}.
}
\examples{
data(cycling_data)
smth_plot(cycling_data, timer.s, power.W, power.smooth.W)  # Default arguments.
}

