% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cycleRdata_S3.R
\name{summary.cycleRdata}
\alias{summary.cycleRdata}
\title{Summary method for cycleRdata class.}
\usage{
\method{summary}{cycleRdata}(object, sRPE = NULL, CP = NULL,
  round_digits = NULL, prettynames = FALSE, .smoothpwr = "power.smooth.W",
  ...)
}
\arguments{
\item{object}{object for which a summary is desired.}

\item{sRPE}{optional; session Rating of Percieved Exertion (value between 1 and 10).}

\item{CP}{optional; Critical Power value (Watts).}

\item{round_digits}{optional; number of digits to round all metrics to.}

\item{prettynames}{self explanatory.}

\item{.smoothpwr}{character string; column name of smoothed power values. For internal use.}

\item{...}{further arguments passed to or from other methods.}
}
\value{
a list object.
}
\description{
Relevant summary metrics for cycling data (class `cycleRdata`).
}
\examples{
data(cycling_data)
summary(cycling_data)

}

