% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Wbal.R
\name{Wbal}
\alias{Wbal}
\title{W' balance.}
\usage{
Wbal(data, time = "timer.s", pwr = "power.W", CP)
}
\arguments{
\item{data}{a data.frame/matrix object with time and power columns.}

\item{time}{character; name of the time (seconds) column in \code{data}.}

\item{pwr}{character; name of the power (watts) column in \code{data}.}

\item{CP}{a critical power value for use in the calculation.}
}
\value{
A numeric vector of W' balance values, in \strong{kilojoules}.
}
\description{
Generate a vector of W' balance values from time and power data. The
underlying algorithm is published in Skiba \emph{et al.} (2012), and can also
be examined in the function source code.
}
\examples{
data(cycling_data)
cycling_data$Wexp.kJ <-
  Wbal(cycling_data, timer.s, power.smooth.W, CP = 300)
}
\references{
Skiba, P. F., W. Chidnok, A. Vanhatalo, and A. M. Jones. Modeling
  the Expenditure and Reconstitution of Work Capacity above Critical Power.
  Med. Sci. Sports Exerc., Vol. 44, No. 8, pp. 1526-1532, 2012.
  \href{http://www.ncbi.nlm.nih.gov/pubmed/22382171}{PubMed link}.
}
\seealso{
\code{\link{Wbal_plots}}.
}

