% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Wbal_plots.R
\name{Wbal_plots}
\alias{Wbal_plots}
\title{W' balance plots.}
\usage{
Wbal_plots(data, x = 1, CP = NULL, n = 3, xlim = NULL)
}
\arguments{
\item{data}{a \strong{formatted} dataset produced by \code{read*()}.}

\item{x}{numeric; \code{1} will plot against time and \code{2} will plot
against distance.}

\item{CP}{a value for critical power annotation.}

\item{n}{how many plots should be created (see details).}

\item{xlim}{passed to \code{\link[graphics]{plot}}.}
}
\value{
a variable number of plots.
}
\description{
Generate three plots that effectively summarise a whole cycling dataset.
}
\details{
When passed a \strong{formatted} dataset, creates a column of three
  plots: \enumerate{\item W' balance. \item Power (raw with 25 sec moving
  averaged overlaid). \item Elevation profile} How many of these plots are
  shown is controlled by the \code{n} argument.
}
\examples{
data(cycling_data)
Wbal_plots(cycling_data, x = 2, CP = 300)
# Show just W' balance.
Wbal_plots(cycling_data, x = 2, n = 1)
# Zoom all plots to 20-40 km.
Wbal_plots(cycling_data, x = 2, CP = 300, xlim = c(20, 40))
}
\seealso{
\code{\link{Wbal}}.
}

