% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zone_index.R
\name{zone_index}
\alias{zone_index}
\title{Index zones.}
\usage{
zone_index(x, zbounds)
}
\arguments{
\item{x}{numeric; values to be "zoned".}

\item{zbounds}{numeric; values for zone boundaries.}
}
\value{
a numeric vector of zone values of the same length as \code{x}. The
  number of zone levels will be \code{length(zbounds) + 1}.
}
\description{
Generate a vector of zone "levels" from an input vector and defined
boundaries.
}
\examples{
data(cycling_data)
cycling_data$zone <- zone_index(cycling_data$power.W, c(100, 200, 300))
}

