% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zone_time.R
\name{zone_time}
\alias{zone_time}
\title{Calculate time in zones.}
\usage{
zone_time(data, column = "power.W", zbounds, pct = FALSE, .string = FALSE)
}
\arguments{
\item{data}{a \strong{formatted} dataset produced by \code{read*()}.}

\item{column}{the column name of the data to which the zone boundaries
relate.}

\item{zbounds}{numeric; zone boundaries.}

\item{pct}{should percentage values be returned?}

\item{.string}{are column name arguments given as character strings? A
backdoor around non-standard evaluation. Mainly for internal use.}
}
\value{
a data frame of zone times.
}
\description{
Given a vector of zone boundaries, sums the time spent in each zone.
}

